Merge the release branch with stable/8.7. This pull request was created
by Zeebe :rocket:.

+++ optimize/backend/src/it/java/io/camunda/optimize/test/it/extension/db/ElasticsearchDatabaseTestService.java
@@ -669,7 +669,10 @@ public class ElasticsearchDatabaseTestService extends DatabaseTestService {
 
   @Override
   public int getNestedDocumentsLimit(final ConfigurationService configurationService) {
-    return configurationService.getElasticSearchConfiguration().getNestedDocumentsLimit();
+    return configurationService
+        .getElasticSearchConfiguration()
+        .getNestedDocumentsLimit()
+        .intValue();
   }
 
   @Override
